refactor(root): collapse 6-way wrangler deploy-script copies into shared libs + shims (#1238)
👤 For humans
Every app carried a byte-identical copy of the two Cloudflare deploy
helpers (~264 lines each set). The logic now lives once at the repo root;
each app keeps a tiny shim at the same path, so package.json cf:* chains
and CI are untouched. Cuts monorepo duplication 15.3% -> 14.8%.

🤖 For agents
- New scripts/lib/sync-wrangler-ids.mjs (exports syncWranglerIds(appDir))
  and scripts/lib/inject-wrangler-env.mjs (exports injectWranglerEnv(appDir))
  — code moved verbatim, appDir parameterized instead of import.meta.url.
- apps/{fanfare,triage,velo} + pocs/{booking-demo,crouton-builder-demo,pins}
  scripts/* replaced by 9-line shims that pass their own app dir.
- packages/crouton-cli/lib/templates/wrangler/* deliberately UNTOUCHED:
  apps scaffolded outside this monorepo need the self-contained copy.
  Logic changes must land in both the lib and the template.
- The moved functions carry fallow-ignore-next-line complexity markers:
  verbatim relocations read as NEW findings to the diff-scoped audit
  (same precedent as scaffold-app.ts); real refactors belong to WS2.
- Verified offline via SYNC_D1_LIST_JSON/SYNC_KV_LIST_JSON dry-runs in all
  6 apps + a fabricated .output fixture for the env injection path.
